提交：登陆，数据库

+++ app/src/main/java/com/brins/lightmusic/model/userlogin/UserAccountBean.kt
@@ -14,7 +14,7 @@ class UserAccountBean {
     @ColumnInfo(name = "ID")
     @PrimaryKey(autoGenerate = true)
     @NonNull
-    var mId: Int = 0
+    var mId: Long = 0
     /**
      * 用户ID
      * */
